Better documentation of bash history config
The goal with these changes to the bash history are to make it so that
we are dynamically writing out the current commands being executed in
any terminal. This is to enable visualization of the last couple of
commands in another session window so that we can have a running log for
classes in SWC. This is done by running something like `watch -n 1 -t
tail -n 3 .bash_history` in another window.